requirements.txt
backports.ssl-match-hostname==3.4.0.2
docker-py==1.1.0
future==0.14.3
python-tutum==0.14.0
requests==2.6.0
six==1.9.0
websocket-client==0.29.0
wsgiref==0.1.2
